+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # Changelog
 
 ## Unreleased
-
+- Added OpenTelemetry Distro and Configurator
+    ([#187](https://github.com/microsoft/ApplicationInsights-Python/pull/187))
 - Initial commit